use FlagStyles.js instead of LockStyles.js for shared styles
